Move RTC functions too seperate files
