+++ Embedded/src/Clock.hpp
@@ -1,0 +1,5 @@
+#include <Arduino.h>
+#include "RTClib.h"
+
+void setupRTC();
+uint32_t getUnixTime();
